refactor(cloudformation): move opt-in/opt-out related tests to opt_test.go
